chore: convert to TypeScript
